refactor(server): extract scrape/crawl cluster into src/server/scrape.js
Largest extraction so far. Pure move, no behavior change. Relocates the
forgeScrape primitive + the whole crawl/extraction cluster (12 functions):
looksLikeSpaShell, _logScrape, _tryUnlocker, _tryScrapingBrowser, forgeScrape,
htmlToMarkdown, getBrandPageContent, rankBrandPages, extractAnchorHrefs,
extractMarkdownLinks, discoverSubpages, _forgeScrapeRateLimited — plus the
_turndown singleton and the SPA_SHELL_RE / _forgeScrapeHits /
FORGE_SCRAPE_RATE_PER_MIN module consts.

Also MOVES the heavy imports (puppeteer-core, @mozilla/readability, jsdom,
turndown) out of server.js into scrape.js — they were used only by this
cluster. scrape.js imports pool (db.js). server.js re-imports the 4 externally
used members (forgeScrape, getBrandPageContent, discoverSubpages,
_forgeScrapeRateLimited) + FORGE_SCRAPE_RATE_PER_MIN (the /api/forge-scrape
route cites it in its 429).

The no-undef gate earned its keep: first pass flagged SPA_SHELL_RE,
_forgeScrapeHits, and FORGE_SCRAPE_RATE_PER_MIN as undefined (consts whose
users moved) before any deploy — fixed by relocating them too. document/window
(Puppeteer page-eval callbacks) are already whitelisted in the gate config.

node --check + no-undef clean, route inventory 213, 41 tests green
(adds test/scrape.test.js for looksLikeSpaShell).
